Reorganized pages and layouts
