+++ app/javascript/components/App.jsx
@@ -1,7 +1,8 @@
 import React, { createContext, useReducer } from 'react';
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
-import Homepage from "./Homepage";
-import Login from "./Login";
+import Homepage from "./pages/Homepage";
+import Login from "./pages/Login";
+import Organization from "./pages/Organization"
 import { initialState, reducer } from "../store/reducer";
 
 
@@ -17,8 +18,10 @@ function App() {
         dispatch
       }}
     >
+    
     <Router>
       <Switch>
+        <Route path="/orgs" component={Organization}/>
         <Route path="/login" component={Login}/>
         <Route path="/" component={Homepage}/>
       </Switch>
